ocr-api/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
httpx==0.27.0
pydantic-settings==2.2.1
fpdf2==2.7.9
Pillow==10.3.0
paddleocr==3.4.1
paddlepaddle==3.3.1
